requirements.txt
cffi==1.1.2
cryptography==0.9.3
enum34==1.0.4
idna==2.0
ipaddress==1.0.14
ndg-httpsclient==0.4.0
pyasn1==0.1.8
pycparser==2.14
pyOpenSSL==0.15.1
PyYAML==3.11
requests==2.7.0
six==1.9.0
